refactor: Article 컴포넌트 헤더 분리,  본문을 Thread 컴포넌트로 변경
