+++ src/components/organisms/ArticleHeader/index.ts
@@ -1,0 +1,1 @@
+export { default as ArticleHeader } from "./ArticleHeader";
